Fixed pull request issues
- replaced direct DOM access with angular's `ViewChild`
- marked subscriptions in the dashboard service with `take(1)` so that they stop listening after one use
- re-did file structure, grouping similar components and services

+++ invictus_dashboard/src/app/dashboard/components/dashboard-task-edit/dashboard-task-edit.component.ts
@@ -1,30 +1,40 @@
 import {
+  AfterViewInit,
   Component,
+  ElementRef,
   EventEmitter,
   Input,
   OnDestroy,
   OnInit,
-  Output
+  Output,
+  ViewChild,
 } from '@angular/core';
 import {
   FormBuilder,
   FormGroup,
   AbstractControl,
   FormArray,
-  Validators
+  Validators,
+  FormControl,
 } from '@angular/forms';
 
-import { DashboardTask, taskPriority } from './dashboard-task';
-import { DashboardService } from './dashboard.service';
-import { NumberValidators } from '../shared/number.validator';
-import { debounceTime } from 'rxjs';
+import { DashboardTask, taskPriority } from '../../services/dashboard-task';
+import { DashboardService } from '../../services/dashboard.service';
+import { NumberValidators } from '../../../shared/number.validator';
+import { debounceTime, tap } from 'rxjs';
+import { MatDialog } from '@angular/material/dialog';
+import { DefaultDialogComponent } from 'src/app/shared/default-dialog.component';
+import { style } from '@angular/animations';
+import { NgStyle } from '@angular/common';
 
 @Component({
   selector: 'dash-dashboard-task-edit',
   templateUrl: './dashboard-task-edit.component.html',
   styleUrls: ['./dashboard-task-edit.component.css'],
 })
-export class DashboardTaskEditComponent implements OnInit, OnDestroy {
+export class DashboardTaskEditComponent
+  implements OnInit, OnDestroy, AfterViewInit
+{
   taskForm: FormGroup;
   apiErrMsg: string = '';
   pageTitle: string = 'Add Task';
@@ -33,6 +43,23 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
   errMsg: { [key: string]: string } = {};
   taskEditModal!: HTMLElement;
   sprintEndDateString!: string;
+  taskFormControls: string[] = [];
+
+  _isDeleteDialog!: boolean;
+  set isDeleteDialog(isDelete: boolean) {
+    this._isDeleteDialog = isDelete;
+    if (!isDelete && this.taskId > 0) {
+      this.onSaveComplete();
+      return;
+    }
+
+    const dialogText: string = this.setDialogText(isDelete);
+    if (isDelete) {
+      this.onDelete(confirm(dialogText));
+    } else {
+      this.onCancel(confirm(dialogText));
+    }
+  }
 
   private validationMsgs = {
     required: 'This field is required.',
@@ -47,21 +74,32 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
   @Input() sprintEndDate: Date;
   @Output() taskChangeEvent: EventEmitter<string> = new EventEmitter<string>();
 
+  @ViewChild('addTaskModal', { read: ElementRef })
+  modalElement!: ElementRef<HTMLElement>;
+  @ViewChild('exitBtn', { read: ElementRef, static: false })
+  exitBtnElement!: ElementRef<HTMLElement>;
 
   constructor(
     private fb: FormBuilder,
-    private dashboardService: DashboardService
+    private dashboardService: DashboardService,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
     this.checkComponentInputs();
     this.buildTaskForm();
+  }
 
+  ngAfterViewInit(): void {
     this.addModalEventListeners();
+    // console.log(this.exitBtnElement);
   }
 
   addModalEventListeners(): void {
-    this.taskEditModal = document.getElementById('addTaskModal');
+    if (!this.modalElement) {
+      return;
+    }
+    this.taskEditModal = this.modalElement.nativeElement;
     this.taskEditModal.addEventListener(
       'shown.bs.modal',
       this.initOnModalOpen.bind(this)
@@ -74,16 +112,16 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
   }
 
   initOnModalOpen(): void {
+    this.apiErrMsg = '';
     this.checkComponentInputs();
     this.buildTaskForm();
   }
 
   checkComponentInputs(): void {
+    this.apiErrMsg = '';
     if (!this.taskId || this.taskId === 0) {
       this.taskId = 0;
-      this.pageTitle = 'Add Task';
     } else {
-      this.pageTitle = 'Editing task with ID: ' + this.taskId;
       this.getTask(this.taskId);
     }
 
@@ -114,8 +152,10 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
       const validation = group.get(i);
       // console.log(i);
       this.errMsg[i] = '';
+      const control = i + controlAppend;
+      this.taskFormControls.push(control);
+      // console.log(this.taskFormControls);
       validation?.valueChanges.pipe(debounceTime(1000)).subscribe((value) => {
-        const control = i + controlAppend;
         // console.log(`${control}: ${value}`);
         this.setMsg(validation, control);
       });
@@ -123,7 +163,6 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
   }
 
   setMsg(c: AbstractControl, control: string): void {
-    //TODO: Show validation messages is touched, dirty, or after the save button has been clicked while untouched
     this.errMsg[control] = '';
     if ((c.touched || c.dirty) && c.errors) {
       console.log(`Error for ${control}: ${Object.keys(c.errors)}`);
@@ -153,7 +192,7 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
 
   addSubTask(): void {
     const numSubTasks = this.subTasks.length;
-    console.log('adding subtask ' + numSubTasks);
+    // console.log('adding subtask ' + numSubTasks);
 
     this.subTasks.push(this.buildTaskGroup());
     const subTaskGroup: FormGroup = (
@@ -169,7 +208,7 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
 
     subTaskGroup.get('id').setValue(subTaskId);
 
-    console.log(`sub-task id: ${subTaskGroup.get('id').value}`);
+    // console.log(`sub-task id: ${subTaskGroup.get('id').value}`);
 
     this.addGroupValueChangeSubscriptions(subTaskGroup, numSubTasks);
   }
@@ -198,7 +237,7 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
     if (this.task.id === 0) {
       this.pageTitle = 'Add Task';
     } else {
-      this.pageTitle = 'Edit ' + this.task.title;
+      this.pageTitle = 'Edit ' + task.title;
     }
 
     this.taskForm.patchValue({
@@ -231,12 +270,12 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
   onSave() {
     if (this.taskForm.valid) {
       if (this.taskForm.dirty) {
-        console.log('Saving...');
+        // console.log('Saving...');
         const t: DashboardTask = { ...this.task, ...this.taskForm.value };
         console.log(t);
 
         if (t.id === 0) {
-          console.log(`creating task: ${t.title}`);
+          // console.log(`creating task: ${t.title}`);
           this.dashboardService.createTask(t).subscribe({
             next: () => {
               // console.log('creating..');
@@ -244,14 +283,20 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
                 this.initSubTasks(this.dashboardService.lastTaskModified);
               }
             },
-            error: (err) => (this.apiErrMsg = err),
+            error: (err) => {
+              this.apiErrMsg = err;
+              console.log(err);
+            },
           });
         } else {
           this.dashboardService.updateTask(t).subscribe({
             next: () => {
               this.initSubTasks(t);
             },
-            error: (err) => (this.apiErrMsg = err),
+            error: (err) => {
+              this.apiErrMsg = err;
+              console.log(err);
+            },
           });
         }
       } else {
@@ -259,20 +304,33 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
       }
     } else {
       this.apiErrMsg = 'Please correct the validation errors.';
+      this.checkAllValidation();
     }
   }
 
-  initSubTasks(task: DashboardTask) {
-    console.log('initiating sub tasks');
+  checkAllValidation(): void {
+    // console.log('checking validation');
+    for (let control in this.taskFormControls) {
+      // console.log(this.taskFormControls[control]);
+      const validation = this.taskForm.get(this.taskFormControls[control]);
+      if (validation) {
+        validation.markAsTouched();
+        validation.updateValueAndValidity();
+      }
+    }
+  }
+
+  initSubTasks(task: DashboardTask): void {
+    // console.log('initiating sub tasks');
     var subTaskAltered: boolean = false;
     for (let subTask in task.subTasks) {
       var subTaskId = task.subTasks[subTask].id;
       if (subTaskId < 1) {
-        console.log(
-          `subtask ${task.subTasks[subTask].title} id: ${
-            task.subTasks[subTask].id
-          } changed to: ${subTaskId + task.id}.`
-        );
+        // console.log(
+        //   `subtask ${task.subTasks[subTask].title} id: ${
+        //     task.subTasks[subTask].id
+        //   } changed to: ${subTaskId + task.id}.`
+        // );
         subTaskId += task.id;
         subTaskAltered = true;
       } else {
@@ -281,44 +339,77 @@ export class DashboardTaskEditComponent implements OnInit, OnDestroy {
     }
 
     if (subTaskAltered) {
-      console.log('sub-tasks added. updating...');
+      // console.log('sub-tasks added. updating...');
       this.dashboardService.updateTask(task).subscribe({
         next: () => this.onSaveComplete(),
-        error: (err) => (this.apiErrMsg = err),
+        error: (err) => {
+          this.apiErrMsg = err;
+          console.log(err);
+        },
       });
     } else {
       this.onSaveComplete();
     }
   }
 
-  onCancel() {
-    //TODO: flesh out this alert menu with back button (modal?)
-    alert(
-      'Warning: leaving the form will clear unsaved data. Are you sure you want to exit?'
-    );
-    this.taskForm.reset();
-  }
-
-  onDelete() {
-    //TODO: flesh out the alert for this too
-    if (!this.task.id || this.task.id === 0) {
-      this.onCancel();
-    } else {
-      this.dashboardService.deleteTask(this.task.id).subscribe({
-        next: () => this.onSaveComplete(),
-        error: (err) => (this.apiErrMsg = err),
-      });
+  onCancel(confirm: boolean) {
+    if (confirm) {
+      this.onSaveComplete();
     }
   }
 
-  onSaveComplete() {
+  onDelete(confirm: boolean) {
+    console.log('delete confirmed: ', confirm);
+    if (!this.task.id || this.task.id === 0) {
+      if (!confirm) {
+        this.onSaveComplete();
+      }
+    } else {
+      if (confirm) {
+        this.dashboardService.deleteTask(this.task.id).subscribe({
+          next: () => this.onSaveComplete(),
+          error: (err) => (this.apiErrMsg = err),
+        });
+      }
+    }
+  }
+
+  setDialogText(isDelete: boolean): string {
+    var modalText: string = '';
+    if (isDelete) {
+      modalText = `Delete ${
+        this.taskForm.get('title').value
+      }? This action cannot be undone.`;
+    } else {
+      modalText =
+        'Leaving the form will clear unsaved data. Are you sure you want to exit?';
+    }
+    return modalText;
+  }
+
+  onSaveComplete(): void {
     this.taskForm.reset();
     this.taskChangeEvent.emit('task');
-    document.getElementById('exitBtn').click();
+    this.exitBtnElement.nativeElement.click();
   }
 
   ngOnDestroy(): void {
     console.log('dashboard destroyed.');
     this.taskEditModal.removeAllListeners();
+  }
+
+  openDialog(bodyText: string, callback: Function): void {
+    const dialogRef = this.dialog.open(DefaultDialogComponent, {
+      width: '300px',
+      data: bodyText,
+      panelClass: 'material-dialog',
+      backdropClass: 'material-dialog',
+    });
+
+    //this will end its own subscription
+    dialogRef.afterClosed().subscribe((result) => {
+      console.log('dialog closed.');
+      callback(result);
+    });
   }
 }
